+++ .idea/misc.xml
@@ -110,7 +110,4 @@
   <component name="ProjectType">
     <option name="id" value="Android" />
   </component>
-  <component name="SvnConfiguration">
-    <configuration>C:\Users\Tony\AppData\Roaming\Subversion</configuration>
-  </component>
 </project>
